// frontend/src/enums/endPoints.enum.ts
export enum EndPoints {
  auth = 'auth',
  register = 'register',
  login = 'login',
  logout = 'logout',
  autoLogin = 'auto-login',
  finance = 'finance',
  transaction = 'transaction',
  getDefaultExpenseCategories = 'get-default-expense-categories',
  getDefaultIncomeCategories = 'get-default-income-categories',
  getDefaultCurrency = 'get-default-currency',
  userPreferences = 'user-prefenrences',
  groupTransaction = 'group-transaction',
  user = 'user',
}